chore: clean up test
